rpm: Clean up error handling for yumdb
Previously this:
went into a chain of GError overwrite warnings because the tree content
didn't have the yumdb data.

Since we're just loading small strings, change the internals to load
it all as a string.  Have the API distinguish between "not found" and
"other error".

Change callers to cleanly throw an error if something goes wrong.